requirements.txt
fastapi>=0.100.0,<1.0.0
pydantic>=2.0.0,<3.0.0
uvicorn==0.22.0
sqlalchemy==2.0.18
PyJWT==2.10.0
psycopg2-binary